backend/requirements.txt
fastapi
uvicorn
python-multipart
pypdf
pdf2image
pytesseract
Pillow
youtube-transcript-api
requests
google-generativeai
pydantic
python-dotenv
sqlalchemy
python-jose[cryptography]
passlib[bcrypt]
bcrypt==4.0.1
moviepy==1.0.3
gTTS
numpy
faiss-cpu
scikit-learn
